requirements.txt
asgiref==3.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2
django-allauth==0.50.0
django-cors-headers==3.13.0
django-filter==21.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
django-sendgrid-v5==1.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
idna==3.3
Markdown==3.3.6
oauthlib==3.2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==1.7.1
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
urllib3==1.26.9
